Test spline functions constraints
Test that `spline::de_casteljau_subdivide` is disabled if there is a
vector and vector space algebra mismatch.

When implemented as a function template, this test is not possible as it
requires a reference to overloaded function outside of allowed contexts.
As a result, `de_casteljau_subdivide` is implemented as a function
object.