Test spline functions constraints
Test that the call operator of `spline::de_casteljau_subdivide` is
disabled if the input types do not form a valid vector space.

When implemented as a function template, this test is not possible as it
requires a reference to an overloaded function outside of allowed
contexts. As a result, `de_casteljau_subdivide` is implemented as a
global function object.